feat : add tokenManager
